rate_limiter/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.8
djangorestframework==3.12.4
idna==3.2
pycodestyle==2.7.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
